tests/wpt/mozilla/tests/bluetooth/getPrimaryServices/services-found.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.two_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryServices())
    .then(services => {
        assert_equals(services.length, 3);
        assert_equals(services[0].uuid, generic_access.uuid);
        assert_equals(services[1].uuid, heart_rate.uuid);
        assert_equals(services[2].uuid, heart_rate.uuid);
        assert_true(services[0].isPrimary);
        assert_true(services[1].isPrimary);
        assert_true(services[2].isPrimary);
    });
}, 'Find all services in a device.');
</script>
